# portal/src/static/openapi.yml
openapi: 3.0.3
info:
  title: Healthcare Services API
  description: This is our Healthcare Services API that let you develop healthcare applications effortlessly.
  version: "1.0"
servers:
  - url: http://localhost:8002
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/opd:
    get:
      tags:
        - "OPD Record Management"
      description: Get all OPD records
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            example: 1
      responses:
        200:
          description: Success
          content:
            application/json:
              example:
                [{"patient_id": "665186afb873a548422f9b79",
                  "opds": [{"ID": "b0a4f529-78b4-4c12-8a1c-50c526384323",
                            "doctor": "Doctor 1",
                            "date": "2024-05-24"},
                            {"ID": "90449f92-ae2a-4de4-80f6-48b90b707f57",
                              "doctor": "Doctor 2",
                              "date": "2024-05-25"}]},
                  {"patient_id": "665186dbb873a548422f9b7b",
                    "opds": [{"ID": "8af925de-8d15-458f-9f25-ea85104760f5",
                              "doctor": "Doctor 1",
                              "date": "2024-05-25"}]}]
  /api/opd/byDate:
    get:
      tags:
        - "OPD Record Management"
      description: Get OPD Records by Date
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            example: "2024-06-04"
        - name: page
          in: query
          required: false
          schema:
            type: integer
            example: 1
      responses:
        200:
          description: Success
          content:
            application/json:
              example:
                [{"patient_id": "665186afb873a548422f9b79",
                  "opds": [{"ID": "b0a4f529-78b4-4c12-8a1c-50c526384323",
                            "doctor": "Doctor 1",
                            "date": "2024-05-24"},
                            {"ID": "90449f92-ae2a-4de4-80f6-48b90b707f57",
                              "doctor": "Doctor 2",
                              "date": "2024-06-04"}]},
                  {"patient_id": "665186dbb873a548422f9b7b",
                    "opds": [{"ID": "8af925de-8d15-458f-9f25-ea85104760f5",
                              "doctor": "Doctor 1",
                              "date": "2024-06-04"}]}]
  /api/opd/{patient_id}:
    get:
      tags:
        - "OPD Record Management"
      description: Get OPD records of a particular Patient
      parameters:
        - name: patient_id
          in: path
          required: true
          schema:
            type: string
            example: 665186afb873a548422f9b79
      responses:
        200:
          description: Success
          content:
            application/json:
              example:
                { "OPD Records": [{"ID": "b0a4f529-78b4-4c12-8a1c-50c526384323",
                                    "doctor": "Doctor 1",
                                    "date": "2024-05-24"},
                                  {"ID": "90449f92-ae2a-4de4-80f6-48b90b707f57",
                                    "doctor": "Doctor 2",
                                    "date": "2024-05-25"}]}

    post:
      tags:
        - "OPD Record Management"
      description: Add OPD Record
      parameters:
        - name: patient_id
          in: path
          required: true
          schema:
            type: string
            example: 665186afb873a548422f9b79
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                doctor:
                  type: string
                  example: <doctor's name here>
                disease_description:
                  type: string
                  example: <disease description here>
                primary_diagnosis:
                  type: string
                  example: <primary diagnosis here>
                prescription:
                  type: string
                  example: <prescription here>
                date:
                  type: string
                  example: "2014-06-04"
                time:
                  type: string
                  example: "10:35:42"
      responses:
        200:
          description: Success
          content:
            application/json:
              example:
                message: "Record added Successfully!"
  /api/opd/{patient_id}/{opd_id}:
    get:
      tags:
        - "OPD Record Management"
      description: Get a specific OPD Record
      parameters:
        - name: patient_id
          in: path
          required: true
          schema:
            type: string
            example: "665186afb873a548422f9b79"
        - name: opd_id
          in: path
          required: true
          schema:
            type: string
            example: "b0a4f529-78b4-4c12-8a1c-50c526384323"
      responses:
        200:
          description: Success
          content:
            application/json:
              example:
                { "OPD Record": {"date": "2024-05-24",
                                "disease_description": "Severe pain and burning in esophagus",
                                "doctor": "Doctor 1",
                                "id": "b0a4f529-78b4-4c12-8a1c-50c526384323",
                                "prescription": "Anti-acidity Medication",
                                "primary_diagnosis": "Peptic Ulcer",
                                "time": "08:18:30"}}

    delete:
      tags:
        - "OPD Record Management"
      description: Delete a specific OPD Record
      parameters:
        - name: patient_id
          in: path
          required: true
          schema:
            type: string
            example: "665186afb873a548422f9b79"
        - name: opd_id
          in: path
          required: true
          schema:
            type: string
            example: "b0a4f529-78b4-4c12-8a1c-50c526384323"
      responses:
        200:
          description: Success
          content:
            application/json:
              example:
                message: Record Deleted Successfully!
  /api/opd/byDate/{patient_id}:
    get:
      tags:
        - "OPD Record Management"
      description: Get a specific Patient's OPD Records by Date
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            example: "2024-06-04"
        - name: patient_id
          in: path
          required: true
          schema:
            type: string
            example: "665186afb873a548422f9b79"
      responses:
        200:
          description: Success
          content:
            application/json:
              example:
                [{ "OPD Records": [{"id": "b0a4f529-78b4-4c12-8a1c-50c526384323",
                                    "doctor": "Doctor 1",
                                    "disease_description": "Sever burning in esophagus",
                                    "prescription": "Anti-acidity Medication",
                                    "primary_diagnosis": "Peptic Ulcer",
                                    "time": "08:18:30",
                                    "date": "2024-06-04"},
                                  {"id": "b0a4f529-78b4-4c12-8a1c-50c526384323",
                                    "doctor": "Doctor 1",
                                    "disease_description": "pain and numbness in both arms",
                                    "prescription": "Muscle relax Medications",
                                    "primary_diagnosis": "Herniated Disc",
                                    "time": "08:18:30",
                                    "date": "2024-06-04"}]}]
...